requirements.txt
biopython>=1.83
pandas>=1.2.2
matplotlib>=3.8.2
requests>=2.31.0

